+++ packages/integration-tests/src/test-e2e/create_amplify.test.ts
@@ -1,5 +1,6 @@
 import { execa } from 'execa';
 import * as fs from 'fs/promises';
+import { existsSync } from 'fs';
 import * as path from 'path';
 import * as os from 'os';
 import { after, afterEach, before, beforeEach, describe, it } from 'node:test';
@@ -8,17 +9,42 @@ import { glob } from 'glob';
 import { testConcurrencyLevel } from './test_concurrency.js';
 import { findBaselineCdkVersion } from '../cdk_version_finder.js';
 import { amplifyAtTag } from '../constants.js';
-import { NpmRegistryController } from '../npm_registry_controller.js';
 
 void describe(
   'create-amplify script',
   { concurrency: testConcurrencyLevel },
   () => {
     let baselineCdkVersion: string;
-    const npmRegistryController = new NpmRegistryController(true);
 
     before(async () => {
-      await npmRegistryController.setUp();
+      // start a local npm proxy and publish the current codebase to the proxy
+      await execa('npm', ['run', 'clean:npm-proxy'], { stdio: 'inherit' });
+      await execa('npm', ['run', 'vend'], { stdio: 'inherit' });
+
+      // nuke the npx cache to ensure we are installing packages from the npm proxy
+      const { stdout } = await execa('npm', ['config', 'get', 'cache']);
+      const npxCacheLocation = path.join(stdout.toString().trim(), '_npx');
+
+      if (existsSync(npxCacheLocation)) {
+        await fs.rm(npxCacheLocation, { recursive: true });
+      }
+
+      // Force 'create-amplify' installation in npx cache by executing help command
+      // before tests run. Otherwise, installing 'create-amplify' concurrently
+      // may lead to race conditions and corrupted npx cache.
+      const output = await execa(
+        'npm',
+        ['create', amplifyAtTag, '--yes', '--', '--help'],
+        {
+          // Command must run outside of 'amplify-backend' workspace.
+          cwd: os.homedir(),
+        }
+      );
+
+      assert.match(output.stdout, /--help/);
+      assert.match(output.stdout, /--version/);
+      assert.match(output.stdout, /Show version number/);
+      assert.match(output.stdout, /--yes/);
 
       // Prefixing with ~. Otherwise, npm is going to install desired version but
       // declare dependency with ^ in package json. So just in case removing that
@@ -28,7 +54,8 @@ void describe(
     });
 
     after(async () => {
-      await npmRegistryController.tearDown();
+      // stop the npm proxy
+      await execa('npm', ['run', 'stop:npm-proxy'], { stdio: 'inherit' });
     });
 
     const startDelayIntervalMS = 1000 * 20; // start tests 20 seconds apart to avoid file hot spots
